+++ view-admin/src/components/common/PrizeResult/index.ts
@@ -1,0 +1,2 @@
+// "PrizeResult"ディレクトリ内のモジュール"PrizeResult.tsx"をdefaultとして再エクスポートする．
+export {default} from "./PrizeResult"
