+++ view-admin/src/components/common/index.ts
@@ -2,4 +2,4 @@ export {default as Header} from "./Header"
 export {default as BingoResult} from "./BingoResult"
 export {default as Button} from "./Button"
 export {default as JudgementModal} from "./JudgementModal"
-
+export {default as PrizeResult} from "./PrizeResult"
