+++ view-admin/src/pages/prizes/prizes.module.css
@@ -1,0 +1,235 @@
+.container {
+    height: 100vh;
+    width: 100vw;
+    overflow: auto;
+    background: linear-gradient(180deg, #d95b7f 0%, #856db2 23.96%, #07033e 100%);
+  }
+  
+  .main {
+    display: flex;
+    align-items: center;
+    margin-right: 1rem;
+  }
+  
+  .buttonContents {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 0.3rem;
+    color: #333;
+    font-weight: 700;
+  }
+  
+  .buttonContents img {
+    color: #fff;
+  }
+  
+  .title {
+    display: flex;
+    justify-content: space-between;
+    color: #fff;
+    gap: 0.3rem;
+    margin-top: 0.8rem;
+    font-size: 1.5rem;
+    font-weight: 700;
+  }
+  
+  .title img {
+    filter: brightness(0) invert(1);
+  }
+  
+  .title_button {
+    display: flex;
+    align-items: center;
+    margin-left: 1rem;
+  }
+  
+  /* 10000円以上の景品 */
+  .grid_01 {
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    grid-gap: 10px;
+    margin: 1rem 1rem 1rem 1rem;
+  }
+  
+  .item_01 {
+    display: grid;
+    place-content: center;
+    place-items: center;
+    width: 22vw;
+    height: 22vw;
+    background: rgba(255, 255, 255, 0.5);
+    border-radius: 0.625rem;
+    position: relative;
+  }
+  
+  .item_01 img {
+    margin-bottom: 5px;
+  }
+  
+  .item_01 p {
+    font-size: 1.5vw;
+    font-weight: 700;
+  }
+  
+  /* 10000円以下の景品 */
+  .grid_02 {
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    grid-gap: 10px;
+    margin: 1rem 1rem 1rem 1rem;
+  }
+  
+  .item_02 {
+    display: grid;
+    place-content: center;
+    place-items: center;
+    width: 22vw;
+    height: 22vw;
+    background: rgba(255, 255, 255, 0.5);
+    border-radius: 0.625rem;
+    position: relative;
+  }
+  
+  .item_02 img {
+    margin-bottom: 5px;
+  }
+  
+  .item_02 p {
+    font-size: 1.5vw;
+    font-weight: 700;
+  }
+  
+  .overlay {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    border-radius: 0.625rem;
+    background-color: rgba(0, 0, 0, 0.3);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  
+  .overlay p {
+    text-align: center;
+    line-height: 1.7;
+    width: 100%;
+    height: 35%;
+    font-size: 2rem;
+    color: #d95b7f;
+    background-color: #07033e;
+  }
+  
+  .overlay_small {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    border-radius: 0.625rem;
+    background-color: rgba(0, 0, 0, 0.3);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  
+  .overlay_small p {
+    text-align: center;
+    line-height: 1.6;
+    width: 100%;
+    height: 35%;
+    font-size: 1.4rem;
+    color: #d95b7f;
+    background-color: #07033e;
+  }
+  
+  .toggle_input {
+    position: absolute;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    z-index: 5;
+    opacity: 0;
+    cursor: pointer;
+  }
+  
+  .toggle_label {
+    width: 85px;
+    height: 42px;
+    background: #fff;
+    border: 3px solid #eee;
+    position: relative;
+    display: inline-block;
+    border-radius: 40px;
+    transition: 0.4s;
+    box-sizing: border-box;
+  }
+  
+  .toggle_label:after {
+    content: "";
+    position: absolute;
+    width: 32px;
+    height: 32px;
+    border-radius: 100%;
+    left: 2px;
+    top: 2px;
+    z-index: 2;
+    background: #eee;
+    transition: 0.4s;
+  }
+  
+  .toggle_input:checked + .toggle_label {
+    border: 3px solid #d95b7f;
+  }
+  
+  .toggle_input:checked + .toggle_label:after {
+    left: 42px;
+    background: #d95b7f;
+  }
+  
+  .toggle_button {
+    position: relative;
+    width: 85px;
+    height: 42px;
+    margin: auto;
+  }
+  
+  .selected .toggle_label {
+    border: 3px solid #d95b7f;
+  }
+  
+  .selected .toggle_label:after {
+    left: 42px;
+    background: #d95b7f;
+  }
+  
+  .search_box {
+    border: 1px solid #07033e;
+    background-color: #ffffff;
+    color: #333;
+    font-weight: 700;
+    padding: 3px 10px;
+    width: 16rem;
+    height: 2.8rem;
+    overflow: hidden;
+    margin-right: 0;
+  }
+  
+  .search_button {
+    cursor: pointer;
+    font-family: FontAwesome;
+    font-size: 1em;
+    background: #07033e;
+    color: #fff;
+    outline: none;
+    width: 3.5em;
+    height: 3rem;
+    display: inline-block;
+    vertical-align: middle;
+    margin-right: 2rem;
+  }
+
